Upgrade to V7.6: Native FP16 Engine with Zero-Overhead Casting 🏎️💨🏆🏁🚀🏁

+++ KudaHitamQuant_full-reasoning.py
@@ -323,7 +323,7 @@ class KudahitamCompressorV2:
     @torch.no_grad()
     def compress(self, states: torch.Tensor, offload: bool = True) -> dict:
         if isinstance(states, (list, tuple)): states = states[0]
-        dev = states.device; shape = [int(v) for v in states.shape]; flat = states.reshape(-1, shape[-1])
+        dev = states.device; shape = [int(v) for v in states.shape]; flat = states.reshape(-1, shape[-1]).half()
         # Random sampling for variance (1024 tokens for speed + accuracy)
         s_size = min(1024, flat.shape[0]); s_idx = torch.randint(0, flat.shape[0], (s_size,), device=dev) if s_size < flat.shape[0] else slice(None)
         
@@ -465,7 +465,7 @@ def main():
                 if k is not None: head_dim = (k[0].shape[-1] if isinstance(k, (tuple, list)) else k.shape[-1]); break
     except Exception as e: print(f"Warning: head_dim detection failed ({e}), using default 256.")
 
-    print(f"Warming up GPU kernels ({model.device})..."); d_k = torch.randn(1, 1, 128, head_dim).to(model.device); d_q = torch.randn(1, 1, 1, head_dim).to(model.device)
+    print(f"Warming up GPU kernels ({model.device})..."); d_k = torch.randn(1, 1, 128, head_dim).to(model.device).half(); d_q = torch.randn(1, 1, 1, head_dim).to(model.device).half()
     for _ in range(5):
         for CC in [KudahitamCompressorV2, KudahitamCompressorGaussian]:
             comp = CC(head_dim, 1, seed=0, device=model.device); c = comp.compress(d_k, offload=False); _ = comp.asymmetric_attention_scores(d_q, c)
